requirements.txt
streamlit>=1.28.0
flask>=2.3.0
google-generativeai>=0.3.0
pandas>=1.5.0
numpy>=1.24.0
matplotlib>=3.6.0
seaborn>=0.12.0
plotly>=5.15.0
ipywidgets>=8.0.0
jupyter>=1.0.0
notebook>=6.5.0
tqdm>=4.64.0
requests>=2.28.0
